Refactor BottomMenu component location
- Move BottomMenu from layout to navigation components
- Update import path in protected layout
- Remove old BottomMenu file
